Create ThumbDemoConfig.py from class in ScrolledThumbnail & ThumbnailCtrl
Add import statements to ScrolledThumbnail and ThumbnailCtrl.
Remove unneeded imports.
